refactor(hub): trocear AutomationsWorkspaceView en hub/automations/ — fase 2 (03/T02)
AutomationsWorkspaceView baja de 1.342 a ~790 líneas extrayendo sin
cambios de comportamiento: automationsShared.ts (tipos de draft,
formatHubDate, EMPTY_DRAFT) y AutomationEditDrawer.tsx (drawer de
crear/editar automatización, ~490). El AutomationsTab (lista + filtros +
import/export) queda en la vista con su API pública intacta; imports
muertos podados.
